refactoring of Controllers
