+++ src/Controller/Account/AddressController.php
@@ -1,68 +1,34 @@
 <?php
 
-namespace App\Controller;
+namespace App\Controller\Account;
 
-use App\Form\PasswordUserType;
 use App\Repository\AddressRepository;
 use Doctrine\ORM\EntityManagerInterface;
-use Doctrine\ORM\Mapping\Entity;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Routing\Attribute\Route;
 use App\Form\AddressUserType;
 use App\Entity\Address;
 
-final class AccountController extends AbstractController
+final class AddressController extends AbstractController
 {
     private EntityManagerInterface $em;
-
     public function __construct(EntityManagerInterface $em)
     {
         // Constructor can be used for dependency injection if needed
         $this->em = $em;
     }   
-        
-   #[Route('/compte', name: 'app_account')]
-   public function index(): Response
-   {
-      return $this->render('account/index.html.twig');
-   }
-
-   #[Route('/compte/modifier-mot-de-passe', name: 'app_account_modify_pwd')]
-   public function password(Request $request, UserPasswordHasherInterface $passwordHasher): Response
-   {
-      $user = $this->getUser();
-      $form = $this->createForm(PasswordUserType::class, $user, [
-         'passwordHasher' => $passwordHasher
-      ]);
-
-      $form->handleRequest($request);
-
-      if ($form->isSubmitted() && $form->isValid()) {
-         $this->em->flush();
-
-         $this->addFlash(
-            'success',
-            'Votre mot de passe est correctement mis à jour'
-         );
-
-      }
-      return $this->render('account/password.html.twig', [
-         'modifyPwd' => $form->createView()
-      ]);
-   }
-   
+       
     #[Route('/compte/adresses', name: 'app_account_addresses')]
-    public function addresses(): Response
+    public function index(): Response
     {
         // Render the addresses in a Twig template
         return $this->render('account/addresses.html.twig');
     }
     
     #[Route('/compte/adresses/delete/{id}', name: 'app_account_address_delete')]
-    public function addressDelete($id, AddressRepository $addressRepository): Response
+    public function delete($id, AddressRepository $addressRepository): Response
     {
         // Fetch the address by ID, ensuring it belongs to the current user
         $address = $addressRepository->findOneById($id);
@@ -86,7 +52,7 @@ final class AccountController extends AbstractController
     }
     
     #[Route('/compte/adresse/ajouter/{id}', name: 'app_account_address_form', defaults: ['id' => null])]
-    public function addressForm(Request $request, $id, AddressRepository $addressRepository): Response
+    public function form(Request $request, $id, AddressRepository $addressRepository): Response
     {
         // If an ID is provided, you might want to fetch the existing address
         if ($id) {
