+++ src/Controller/Account/HomeController.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Controller\Account;
+
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Attribute\Route;
+
+final class HomeController extends AbstractController
+{
+    #[Route('/compte', name: 'app_account')]
+    public function index(): Response
+    {
+        return $this->render('account/index.html.twig');
+    }
+}
